feito algumas altercoes de servico e testes, e criado o ProductIntegrationTest

+++ DonaLaura/DonaLaura.Domain/Features/Products/IProductRepository.cs
@@ -8,6 +8,6 @@ namespace DonaLaura.Domain.Features.Products
 
         Product GetByName(string name);
 
-        bool IsTiedTo(int id);
+        bool IsTiedTo(long id);
     }
 }
